moved all the relevant files to client_side_encryption folder
