+++ spec/integration/client_side_encryption/auto_encryption_command_monitoring_spec.rb
