+++ spec/integration/client_side_encryption/auto_encryption_old_wire_version_spec.rb
