+++ spec/integration/client_side_encryption/auto_encryption_reconnect_spec.rb
