+++ spec/integration/client_side_encryption/auto_encryption_spec.rb
