+++ spec/integration/client_side_encryption/explicit_encryption_spec.rb
